refactor(apps/web): 컴포넌트 분리 및 자잘한 수정
